Add first pass at a simple replacement for Celery, that requires no extra setup on the part of the user.
Instead of requiring celeryd to be running, and on the assumption that we will mostly be running a few very-long-running jobs rather than thousands of shorter jobs, we use a daemonizing wrapper, installed as `openspendingetld`, which can run jobs in the Pylons environment in the background, writing output to a log file, which can in turn be read by the Pylons frontend.

See the file docstring in etl/command/daemon.py for more information.