Don't add create "dimensions" for "measures"
This change ensures that "dimensions" are only created for fields that
are truly dimensions and not measures. In OpenSpending terms, this means
that we only create dimension objects for entities and classifiers, and
*not* for values.

The motivation for this change comes from cube computation, where we
don't want to be computing aggregates over "measures" such as "amount",
which typically take values from a continuum rather than a discrete set.